# Conflicts:
#	Hive Armada/Assets/Prefabs/Ship/FixedShip.prefab
#	Hive Armada/Assets/Scenes/Menu Room.unity
#	Hive Armada/Assets/Scenes/Wave Room.unity

+++ Hive Armada/Assets/Scripts/Enemies/Boss.cs
@@ -22,7 +22,7 @@ namespace Hive.Armada.Enemies
         /// <summary>
         /// Type identifier for object pooling purposes
         /// </summary>
-        private int projectileTypeIdentifier;
+        private short projectileTypeIdentifier;
 
         /// <summary>
         /// Structure resposible for tracking the positions for which bullets
